+++ lib/state.dart
@@ -1,0 +1,35 @@
+/// Copyright (C) 2018-2022 Jason C.H
+///
+/// This library is free software; you can redistribute it and/or
+/// modify it under the terms of the GNU Lesser General Public
+/// License as published by the Free Software Foundation; either
+/// version 2.1 of the License, or (at your option) any later version.
+///
+/// This library is distributed in the hope that it will be useful,
+/// but WITHOUT ANY WARRANTY; without even the implied warranty of
+/// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+/// Lesser General Public License for more details.
+
+// ignore_for_file: constant_identifier_names
+
+/// The generic VPN state for all platforms.
+enum FlutterVpnState {
+  disconnected,
+  connecting,
+  connected,
+  disconnecting,
+  error,
+}
+
+/// The error state from `VpnStateService`.
+/// Only available for Android.
+enum CharonErrorState {
+  NO_ERROR,
+  AUTH_FAILED,
+  PEER_AUTH_FAILED,
+  LOOKUP_FAILED,
+  UNREACHABLE,
+  GENERIC_ERROR,
+  PASSWORD_MISSING,
+  CERTIFICATE_UNAVAILABLE,
+}